[refactor] 북마크 삭제
