+++ src/main/java/potatoes/server/travel/domain/query/BookmarkQuery.java
@@ -1,10 +1,13 @@
 package potatoes.server.travel.domain.query;
 
+import java.util.Optional;
+
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Component;
 
 import lombok.RequiredArgsConstructor;
+import potatoes.server.travel.bookmark.entity.Bookmark;
 import potatoes.server.travel.bookmark.repository.BookmarkRepository;
 import potatoes.server.travel.dto.GetMyTravelResponse;
 
@@ -13,6 +16,20 @@ import potatoes.server.travel.dto.GetMyTravelResponse;
 public class BookmarkQuery {
 
 	private final BookmarkRepository bookmarkRepository;
+
+	public Bookmark getBookmark(Long userId, Long travelId) {
+		return bookmarkRepository.findByUserIdAndTravelId(userId, travelId);
+	}
+
+	public Boolean isUserParticipating(Optional<Long> userId, Long travelId) {
+		return userId
+			.map(uid -> bookmarkRepository.existsByUserIdAndTravelId(uid, travelId))
+			.orElse(null);
+	}
+
+	public Boolean isUserParticipating(Long userId, Long travelId) {
+		return isUserParticipating(Optional.of(userId), travelId);
+	}
 
 	public Page<GetMyTravelResponse> findTravelsByBookmark(Pageable pageable, Long userId) {
 		return bookmarkRepository.findMyTravelsByBookmark(pageable, userId);
